+++ gakumas-tools/simulator/CHANGELOG.md
@@ -16,6 +16,17 @@ Dates are based on Eastern Time Zone.
 
 
 
+
+
+## 2025-08-29
+
+- Update for Miracle Na Now Mao and Sumika
+
+## 2025-08-28
+
+- Added preview of na now mao
+- Apply motivation buffs to motivation
+
 ## 2025-08-25
 
 - Fix lock stance turns being applied to full power ending